+++ src/lib/index.ts
@@ -1,12 +1,11 @@
-import { arrayifyInput, InvalidInputError, IsInputValid, mergeOptions, Options } from './helpers';
 import { Killer } from './killer';
 
-export async function killPortProcess(input: any, options: Options = {}) {
-	try {
-		const validInput = IsInputValid(input);
+type Input = number | number[] | string | string[];
 
-		if (!validInput) {
-			throw new InvalidInputError('Invalid input', input);
+export async function killPortProcess(input: Input, options: Options = {}) {
+	try {
+		if (isNullOrUndefined(input)) {
+			throw new Error('No port(s) found');
 		}
 
 		const mergedOptions = mergeOptions(options);
@@ -19,4 +18,25 @@ export async function killPortProcess(input: any, options: Options = {}) {
 	} catch (error) {
 		throw error;
 	}
+}
+
+export function isNullOrUndefined(input: any) {
+	if (input === undefined || input === null) {
+		return true;
+	}
+	return false;
+}
+
+export function arrayifyInput(input) {
+	return Array.isArray(input) ? input : [input];
+}
+
+export interface Options {
+	[key: string]: string;
+}
+
+export function mergeOptions(options: Options) {
+	const defaultOptions = {};
+
+	return { ...defaultOptions, ...options };
 }
